#182 Fix PhpStan Error

+++ src/PrestashopConsole/Command/Webservice/DeleteKeyCommand.php
@@ -77,7 +77,11 @@ class DeleteKeyCommand extends Command
                 WHERE `key`= '" . pSQL($apiKey) . "' 
         ");
 
-        $webserviceKey = new WebserviceKey($idKey);
+        if (false === $idKey) {
+            $output->writeln('<error>The webservice key does not exists</error>');
+            return self::RESPONSE_ERROR;
+        }
+        $webserviceKey = new WebserviceKey((int)$idKey);
         try {
             $webserviceKey->delete();
         } catch (PrestaShopException $e) {
